backend/requirements.txt
annotated-types==0.7.0
anyio
certifi==2025.4.26
click==8.1.8
dnspython
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe
mdurl==0.1.2
pydantic
pydantic_core
Pygments==2.19.1
python-dotenv
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.6
shellingham==1.5.4
sniffio==1.3.1
starlette
typer==0.15.3
typing_extensions
uvicorn
uvloop==0.21.0
watchfiles
websockets
motor
python-dotenv
python-jose
passlib